Fixed some more bugs
Fixed a bug in ObjCheckLeftWallDist where there was a missing instruction. It was documented in the disassembly, but commented out for some reason. Not sure why.

Fixed a bug where spindashing next to a moving object would take a character out of their spindash animation and force the walking animation instead.
Also fixed an animation issue with Tails. Both described here:
https://forums.sonicretro.org/index.php?threads/some-changes-and-fixes-for-sonic-2.29029/page-11#post-876557